+++ crates/motsu/src/context.rs
@@ -4,9 +4,9 @@ use core::fmt::Debug;
 use std::{
     cell::Cell,
     collections::HashMap,
-    hash::Hash,
     ops::{Deref, DerefMut},
     ptr, slice,
+    sync::LazyLock,
     thread::ThreadId,
 };
 
@@ -15,17 +15,12 @@ use alloy_signer_local::PrivateKeySigner;
 use alloy_sol_types::{SolEvent, Word};
 use dashmap::{mapref::one::RefMut, DashMap};
 use k256::ecdsa::SigningKey;
-use once_cell::sync::Lazy;
 use stylus_sdk::{
-    host::{WasmVM, VM},
-    keccak_const::Keccak256,
-    prelude::StorageType,
-    types::AddressVM,
-    ArbResult,
+    keccak_const::Keccak256, prelude::StorageType, types::AddressVM, ArbResult,
 };
 
 use crate::{
-    router::{VMRouter, VMRouterContext},
+    router::{Router, VMRouter},
     storage_access::AccessStorage,
 };
 
@@ -33,14 +28,13 @@ use crate::{
 ///
 /// A global mutable key-value store that allows concurrent access.
 ///
-/// The key is the test [`VMContext`], an id of the test thread.
+/// The key is the test [`VM`], an id of the test thread.
 ///
-/// The value is the [`VMContextStorage`], a storage of the test case.
+/// The value is the [`VMStorage`], a storage of the test case.
 ///
-/// NOTE: The [`VMContext::storage`] will panic on lock, when the same key is
+/// NOTE: The [`VM::storage`] will panic on lock, when the same key is
 /// accessed twice from the same thread.
-static MOTSU_VM: Lazy<DashMap<VMContext, VMContextStorage>> =
-    Lazy::new(DashMap::new);
+static MOTSU_VM: LazyLock<DashMap<VM, VMStorage>> = LazyLock::new(DashMap::new);
 
 // TODO: remove this after we can enable the `stylus-test` feature, which should
 // happen after we refactor `motsu` to implement a mock
@@ -53,14 +47,14 @@ pub(crate) const DEFAULT_CHAIN_ID: u64 = 42161;
 /// Context of Motsu test VM associated with the current test thread.
 #[allow(clippy::module_name_repetitions)]
 #[derive(Hash, Eq, PartialEq, Copy, Clone)]
-pub struct VMContext {
+pub struct VM {
     thread_id: ThreadId,
 }
 
-impl VMContext {
-    /// Get test context associated with the current test thread.
+impl VM {
+    /// Get test `VM` associated with the current test thread.
     #[must_use]
-    pub fn current() -> Self {
+    pub fn context() -> Self {
         Self { thread_id: std::thread::current().id() }
     }
 
@@ -154,7 +148,7 @@ impl VMContext {
 
     /// Initialise contract's storage for the current test thread and
     /// `contract_address`.
-    fn init_storage<ST: StorageType + VMRouter + 'static>(
+    fn init_storage<ST: StorageType + Router + 'static>(
         self,
         contract_address: Address,
     ) {
@@ -172,10 +166,10 @@ impl VMContext {
         self.router(contract_address).init_storage::<ST>();
     }
 
-    /// Reset storage for the current [`VMContext`] and `contract_address`.
+    /// Reset storage for the current [`VM`] and `contract_address`.
     ///
     /// If all test contracts are removed, flush storage for the current
-    /// test [`VMContext`].
+    /// test [`VM`].
     fn reset_storage(self, contract_address: Address) {
         let mut storage = self.storage();
         storage.persistent.contracts.remove(&contract_address);
@@ -197,30 +191,14 @@ impl VMContext {
         address: *const u8,
         calldata: *const u8,
         calldata_len: usize,
-        return_data_size: *mut usize,
-    ) -> u8 {
-        let address = read_address(address);
-        let (selector, input) = decode_calldata(calldata, calldata_len);
-
-        let result = self.call_contract(address, selector, &input, None);
-        self.process_arb_result_raw(result, return_data_size)
-    }
-
-    /// Call the contract at raw `address` with the given raw `calldata` and
-    /// `value`.
-    pub(crate) unsafe fn call_contract_with_value_raw(
-        self,
-        address: *const u8,
-        calldata: *const u8,
-        calldata_len: usize,
         value: *const u8,
         return_data_size: *mut usize,
     ) -> u8 {
         let address = read_address(address);
         let value = read_u256(value);
-        let (selector, input) = decode_calldata(calldata, calldata_len);
+        let calldata = slice::from_raw_parts(calldata, calldata_len);
 
-        let result = self.call_contract(address, selector, &input, Some(value));
+        let result = self.call_contract(address, calldata, Some(value));
         self.process_arb_result_raw(result, return_data_size)
     }
 
@@ -250,8 +228,7 @@ impl VMContext {
     fn call_contract(
         self,
         contract_address: Address,
-        selector: u32,
-        input: &[u8],
+        calldata: &[u8],
         value: Option<U256>,
     ) -> ArbResult {
         // Set the caller contract as message sender and callee contract as
@@ -276,15 +253,19 @@ impl VMContext {
         // Call external contract.
         let result = self
             .router(contract_address)
-            .route(selector, input)
-            .unwrap_or_else(|| {
-                panic!("selector not found - selector is {selector}")
+            .route(calldata.to_vec())
+            .map_err(|e| {
+                // If the call was unsuccessful, we should restore the data.
+                self.storage().persistent.restore_from(backup);
+                // The nested `router_entrypoint` call returns `Err(Vec::new())` when no function
+                // was found for the selector and no fallback is present.
+                if e.is_empty() {
+                    let selector = decode_selector(calldata);
+                    format!("function not found for selector '{selector}' and no fallback defined").as_bytes().to_vec()
+                } else {
+                    e
+                }
             });
-
-        // If the call was unsuccessful, we should restore the data.
-        if result.is_err() {
-            self.storage().persistent.restore_from(backup);
-        }
 
         // Set the previous message sender and contract address back.
         _ = self.replace_contract_address(previous_contract_address);
@@ -552,13 +533,13 @@ impl VMContext {
     }
 
     /// Get reference to the storage for the current test thread.
-    fn storage(self) -> RefMut<'static, VMContext, VMContextStorage> {
+    fn storage(self) -> RefMut<'static, VM, VMStorage> {
         MOTSU_VM.access_storage(&self)
     }
 
     /// Get router for the contract at `address`.
-    fn router(self, address: Address) -> VMRouterContext {
-        VMRouterContext::new(self.thread_id, address)
+    fn router(self, address: Address) -> VMRouter {
+        VMRouter::new(self.thread_id, address)
     }
 
     /// Get the current chain ID.
@@ -609,21 +590,15 @@ pub(crate) unsafe fn write_u256(ptr: *mut u8, value: U256) {
     ptr::copy(bytes.as_ptr(), ptr, 32);
 }
 
-/// Decode the selector as [`u32`], and function input as [`Vec<u8>`] from the
-/// raw pointer.
-unsafe fn decode_calldata(
-    calldata: *const u8,
-    calldata_len: usize,
-) -> (u32, Vec<u8>) {
-    let calldata = slice::from_raw_parts(calldata, calldata_len);
+/// Decode the selector as [`u32`] from the raw pointer to the calldata.
+fn decode_selector(calldata: &[u8]) -> u32 {
     let selector =
         u32::from_be_bytes(TryInto::try_into(&calldata[..4]).unwrap());
-    let input = calldata[4..].to_vec();
-    (selector, input)
+    selector
 }
 
 /// Main storage for Motsu test VM.
-struct VMContextStorage {
+struct VMStorage {
     /// Address of the message sender.
     msg_sender: Option<Address>,
     /// The ETH value in wei sent to the program.
@@ -642,7 +617,7 @@ struct VMContextStorage {
     chain_id: u64,
 }
 
-impl Default for VMContextStorage {
+impl Default for VMStorage {
     fn default() -> Self {
         Self {
             msg_sender: None,
@@ -758,10 +733,9 @@ pub struct ContractCall<'a, ST: StorageType> {
 impl<ST: StorageType> ContractCall<'_, ST> {
     /// Preset the call parameters.
     fn set_call_params(&self) {
-        _ = VMContext::current().replace_optional_msg_value(self.msg_value);
-        _ = VMContext::current().replace_msg_sender(self.msg_sender);
-        _ = VMContext::current()
-            .replace_contract_address(self.contract_ref.address);
+        VM::context().replace_optional_msg_value(self.msg_value);
+        VM::context().replace_msg_sender(self.msg_sender);
+        VM::context().replace_contract_address(self.contract_ref.address);
     }
 
     /// Invalidate the storage cache of stylus contract [`StorageType`], by
@@ -778,7 +752,7 @@ impl<ST: StorageType> Deref for ContractCall<'_, ST> {
 
     #[inline]
     fn deref(&self) -> &Self::Target {
-        VMContext::current().backup();
+        VM::context().backup();
 
         // Set parameters for call such as `msg_sender`, `contract_address`,
         // `msg_value`.
@@ -786,7 +760,7 @@ impl<ST: StorageType> Deref for ContractCall<'_, ST> {
 
         // Transfer value (if any) from the `msg_sender` to `contract_address`,
         // that was set on the previous step.
-        VMContext::current().transfer_value();
+        VM::context().transfer_value();
 
         // SAFETY: We don't use `ST` contract type as intended by rust.
         // We don't care about any state it has in any property.
@@ -801,7 +775,7 @@ impl<ST: StorageType> Deref for ContractCall<'_, ST> {
 impl<ST: StorageType> DerefMut for ContractCall<'_, ST> {
     #[inline]
     fn deref_mut(&mut self) -> &mut Self::Target {
-        VMContext::current().backup();
+        VM::context().backup();
 
         // Set parameters for call such as `msg_sender`, `contract_address`,
         // `msg_value`.
@@ -809,7 +783,7 @@ impl<ST: StorageType> DerefMut for ContractCall<'_, ST> {
 
         // Transfer value (if any) from the `msg_sender` to `contract_address`,
         // that was set on the previous step.
-        VMContext::current().transfer_value();
+        VM::context().transfer_value();
 
         self.invalidate_storage_type_cache();
         self.storage.get_mut()
@@ -824,17 +798,17 @@ pub struct Contract<ST: StorageType> {
 
 impl<ST: StorageType> Drop for Contract<ST> {
     fn drop(&mut self) {
-        VMContext::current().reset_storage(self.address);
+        VM::context().reset_storage(self.address);
     }
 }
 
-impl<ST: StorageType + VMRouter + 'static> Default for Contract<ST> {
+impl<ST: StorageType + Router + 'static> Default for Contract<ST> {
     fn default() -> Self {
         Contract::new_at(Address::default())
     }
 }
 
-impl<ST: StorageType + VMRouter + 'static> Contract<ST> {
+impl<ST: StorageType + Router + 'static> Contract<ST> {
     /// Create a new contract with default storage on the random address.
     #[must_use]
     pub fn new() -> Self {
@@ -844,7 +818,7 @@ impl<ST: StorageType + VMRouter + 'static> Contract<ST> {
     /// Create a new contract with the given `address`.
     #[must_use]
     pub fn new_at(address: Address) -> Self {
-        VMContext::current().init_storage::<ST>(address);
+        VM::context().init_storage::<ST>(address);
 
         Self { phantom: ::core::marker::PhantomData, address }
     }
@@ -902,7 +876,7 @@ impl<ST: StorageType + VMRouter + 'static> Contract<ST> {
 
     /// Check if the `event` was emitted, by the contract `self`.
     pub fn emitted<E: SolEvent>(&self, event: &E) -> bool {
-        VMContext::current().emitted_for(&self.address, event)
+        VM::context().emitted_for(&self.address, event)
     }
 
     /// Assert that the `event` was emitted, by the contract `self`.
@@ -915,7 +889,7 @@ impl<ST: StorageType + VMRouter + 'static> Contract<ST> {
     ///   any) in the error message.
     #[track_caller]
     pub fn assert_emitted<E: SolEvent + Debug>(&self, event: &E) {
-        let context = VMContext::current();
+        let context = VM::context();
         if context.emitted_for(&self.address, event) {
             return;
         }
@@ -936,19 +910,20 @@ impl<ST: StorageType + VMRouter + 'static> Contract<ST> {
 /// Create a default [`StorageType`] `ST` type with at [`U256::ZERO`] slot and
 /// `0` offset.
 pub(crate) fn create_default_storage_type<ST: StorageType>() -> ST {
-    #[cfg(target_arch = "wasm32")]
     unsafe {
-        ST::new(U256::ZERO, 0, VM(WasmVM {}))
-    }
-    #[cfg(not(target_arch = "wasm32"))]
-    unsafe {
-        ST::new(U256::ZERO, 0, VM { host: Box::new(WasmVM {}) })
+        ST::new(
+            U256::ZERO,
+            0,
+            stylus_sdk::host::VM {
+                host: Box::new(stylus_sdk::host::WasmVM {}),
+            },
+        )
     }
 }
 
 /// Account that can be used to interact with contracts in test environments.
 /// Used to interact with and sign transactions for contracts.
-#[derive(Clone, Copy)]
+#[derive(Clone, Copy, PartialEq, Eq, Debug)]
 pub struct Account {
     address: Address,
     private_key: B256,
@@ -957,6 +932,36 @@ pub struct Account {
 impl From<Account> for Address {
     fn from(value: Account) -> Self {
         value.address
+    }
+}
+
+impl From<&Account> for Address {
+    fn from(value: &Account) -> Self {
+        value.address
+    }
+}
+
+impl From<PrivateKeySigner> for Account {
+    fn from(value: PrivateKeySigner) -> Self {
+        Self { address: value.address(), private_key: value.to_bytes() }
+    }
+}
+
+impl From<&PrivateKeySigner> for Account {
+    fn from(value: &PrivateKeySigner) -> Self {
+        Self { address: value.address(), private_key: value.to_bytes() }
+    }
+}
+
+impl From<Account> for PrivateKeySigner {
+    fn from(value: Account) -> Self {
+        value.signer()
+    }
+}
+
+impl From<&Account> for PrivateKeySigner {
+    fn from(value: &Account) -> Self {
+        value.signer()
     }
 }
 
@@ -979,8 +984,9 @@ impl Account {
     #[must_use]
     pub fn from_seed_slice(seed: &[u8]) -> Self {
         let private_key_bytes = Keccak256::new().update(seed).finalize();
-        let address = create_signer(&private_key_bytes).address();
-        let private_key = private_key_bytes.into();
+        let signer = create_signer(&private_key_bytes);
+        let address = signer.address();
+        let private_key = private_key_bytes.into(); // same as signer.to_bytes()
         Self { address, private_key }
     }
 
@@ -1019,7 +1025,7 @@ pub trait Funding {
 
 impl Funding for Address {
     fn fund(&self, value: U256) {
-        VMContext::current().add_assign_balance(*self, value);
+        VM::context().add_assign_balance(*self, value);
     }
 }
 
@@ -1029,7 +1035,7 @@ impl Funding for Account {
     }
 }
 
-impl<ST: StorageType + VMRouter + 'static> Funding for Contract<ST> {
+impl<ST: StorageType + Router + 'static> Funding for Contract<ST> {
     fn fund(&self, value: U256) {
         self.address().fund(value);
     }
@@ -1047,7 +1053,7 @@ impl Balance for Account {
     }
 }
 
-impl<ST: StorageType + VMRouter + 'static> Balance for Contract<ST> {
+impl<ST: StorageType + Router + 'static> Balance for Contract<ST> {
     fn balance(&self) -> U256 {
         self.address().balance()
     }
@@ -1070,7 +1076,7 @@ impl FromTag for Account {
     /// Also registers the tag in the test context for debugging purposes.
     fn from_tag(tag: &str) -> Self {
         let account = Account::from_seed(tag);
-        VMContext::current().set_tag(account.address(), tag.to_string());
+        VM::context().set_tag(account.address(), tag.to_string());
         account
     }
 }
@@ -1082,12 +1088,12 @@ impl FromTag for Address {
     fn from_tag(tag: &str) -> Self {
         let hash = Keccak256::new().update(tag.as_bytes()).finalize();
         let address = Address::from_slice(&hash[..20]);
-        VMContext::current().set_tag(address, tag.to_string());
+        VM::context().set_tag(address, tag.to_string());
         address
     }
 }
 
-impl<ST: StorageType + VMRouter + 'static> FromTag for Contract<ST> {
+impl<ST: StorageType + Router + 'static> FromTag for Contract<ST> {
     /// Creates a contract at an address derived from the tag string.
     ///
     /// This allows deploying contracts to deterministic addresses for testing.
@@ -1103,9 +1109,13 @@ mod tests {
     use stylus_sdk::prelude::*;
 
     use super::{Account, Address, Contract, FromTag};
-    use crate::context::VMContext;
+    use crate::context::VM;
 
     mod account {
+        use std::ops::Deref;
+
+        use alloy_signer_local::PrivateKeySigner;
+
         use super::*;
 
         #[test]
@@ -1150,6 +1160,42 @@ mod tests {
             let signer = account.signer();
             assert_eq!(account.address(), signer.address());
         }
+
+        #[test]
+        fn account_and_its_signer_have_same_private_keys() {
+            let account = Account::from_seed("seed");
+            let signer = account.signer();
+            let signing_key = signer.to_bytes();
+
+            assert_eq!(account.private_key, signing_key);
+        }
+
+        #[test]
+        fn account_signer_and_back_returns_same_account() {
+            let old_account = Account::from_seed("seed");
+            let signer = old_account.signer();
+            let new_account = signer.into();
+            assert_eq!(old_account, new_account);
+        }
+
+        #[test]
+        fn account_into_signer_and_back_returns_same_account() {
+            let old_account = Account::from_seed("seed");
+            let signer: PrivateKeySigner = old_account.into();
+            let new_account = signer.into();
+            assert_eq!(old_account, new_account);
+        }
+
+        #[test]
+        fn account_ref_into_address() {
+            let account = &Account::random();
+            let address: Address = account.into();
+            assert_eq!(account.address(), address);
+
+            let account = &mut Account::random();
+            let address: Address = account.deref().into();
+            assert_eq!(account.address(), address);
+        }
     }
 
     mod from_tag {
@@ -1163,13 +1209,10 @@ mod tests {
                 address!("0x6e12d8c87503d4287c294f2fdef96acd9dff6bd2");
 
             let account = Account::from_tag(&tag);
+
             assert_eq!(expected_private_key, account.private_key);
             assert_eq!(expected_address, account.address());
-
-            assert_eq!(
-                Some(tag),
-                VMContext::current().get_tag(account.address())
-            );
+            assert_eq!(Some(tag), VM::context().get_tag(account.address()));
         }
 
         #[test]
@@ -1179,9 +1222,9 @@ mod tests {
                 address!("0x9c0257114eb9399a2985f8e75dad7600c5d89fe3");
 
             let address = Address::from_tag(&tag);
-            assert_eq!(expected_address, address);
 
-            assert_eq!(Some(tag), VMContext::current().get_tag(address));
+            assert_eq!(expected_address, address);
+            assert_eq!(Some(tag), VM::context().get_tag(address));
         }
 
         #[storage]
@@ -1199,12 +1242,28 @@ mod tests {
                 address!("0x7f6dd79f0020bee2024a097aaa5d32ab7ca31126");
 
             let contract = Contract::<SomeContract>::from_tag(&tag);
-            assert_eq!(expected_address, contract.address());
 
-            assert_eq!(
-                Some(tag),
-                VMContext::current().get_tag(contract.address())
-            );
+            assert_eq!(expected_address, contract.address());
+            assert_eq!(Some(tag), VM::context().get_tag(contract.address()));
+        }
+
+        #[test]
+        fn tag_maps_to_different_address_for_account() {
+            let tag = "tag";
+
+            let address = Address::from_tag(tag);
+            let account = Account::from_tag(tag);
+            let contract = Contract::<SomeContract>::from_tag(tag);
+
+            // account uses a different algorithm for deriving the address
+            assert_eq!(address, contract.address());
+            assert_ne!(address, account.address());
+
+            // all addresses still map to the same tag
+            let tag = Some(tag.to_owned());
+            assert_eq!(tag, VM::context().get_tag(address));
+            assert_eq!(tag, VM::context().get_tag(account.address()));
+            assert_eq!(tag, VM::context().get_tag(contract.address()));
         }
     }
 }